Support alt credential scope and host skip
Instead of defaulting to using the hostname as the credential scope (3rd
field in the credential string, a.k.a. service name), allow us to
explicitly set it so we can use custom domains with AWS API gateway.

Also allow explicit control over whether or not the Host header is
signed. The AWS4 HMAC standard requires it be signed, but some
non-standard implementations do not sign it.